requirements.txt
Django>=2.2
djangorestframework==3.14
markdown
django-filter
psycopg2
drf_spectacular
django-rest-swagger
djangorestframework-simplejwt
